+++ src/pages/index.ts
@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-import { GenerateToken } from './GenerateToken';
+import { ChangelogPage } from './Changelog';
 import { RequirementsPage } from './Requirements';
 import { WelcomePage } from './Welcome';
 
 export {
-  GenerateToken,
+  ChangelogPage,
   RequirementsPage,
   WelcomePage,
 };